requirements.txt
asttokens==3.0.1
bcrypt==5.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
executing==2.2.1
idna==3.11
iniconfig==2.3.0
invoke==2.2.1
ipdb==0.13.13
ipython==9.8.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
junos-eznc==2.7.5
librt==0.7.7
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mypy==1.19.1
mypy-extensions==1.1.0
napalm==5.1.0
ncclient==0.7.0
netaddr==1.3.0
netmiko==4.6.0
netutils==1.15.2
nornir==3.5.0
nornir-jinja2==0.2.0
nornir-netmiko==1.0.1
nornir-rich @ git+https://github.com/sjhloco/nornir_rich@cf2674259c24d001b71831482840b9ebd97caf21
nornir-utils==0.2.0
nornir-validate==0.3.0
ntc-templates @ git+https://github.com/networktocode/ntc-templates.git@286039ef3a4c9b5899ef35eb7e662f44c7328576
packaging==25.0
paramiko==4.0.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pluggy==1.6.0
prompt-toolkit==3.0.52
ptyprocess==0.7.0
pure-eval==0.2.3
pyaml==25.7.0
pycparser==2.23
pyeapi==1.0.4
pygments==2.19.2
pynacl==1.6.2
pyparsing==3.3.1
pyserial==3.5
pytest==9.0.2
pyyaml==6.0.3
requests==2.32.5
rich==13.9.4
ruamel-yaml==0.18.16
ruff==0.14.10
scp==0.15.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
textfsm==2.1.0
traitlets==5.14.3
transitions==0.9.3
ttp==0.10.0
ttp-templates==0.3.7
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
urllib3==2.6.2
wcwidth==0.2.14
yamlordereddictloader==0.4.2
